Struct pallet_nft::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn collection_owner( collection_id: T::NftCollectionId ) -> Option<T::AccountId>

pub fn owner( collection_id: T::NftCollectionId, item_id: T::NftItemId ) -> Option<T::AccountId>

pub fn do_create_collection( owner: T::AccountId, collection_id: T::NftCollectionId, created_by: Acc, metadata: BoundedVecOfUnq<T> ) -> DispatchResult

pub fn do_mint( owner: T::AccountId, collection_id: T::NftCollectionId, item_id: T::NftItemId, metadata: BoundedVecOfUnq<T> ) -> DispatchResult

pub fn set_metadata( owner: T::AccountId, collection_id: T::NftCollectionId, item_id: T::NftItemId, metadata: BoundedVecOfUnq<T> ) -> DispatchResult

pub fn do_transfer( collection_id: T::NftCollectionId, item_id: T::NftItemId, from: T::AccountId, to: T::AccountId ) -> DispatchResult

pub fn do_burn( owner: T::AccountId, collection_id: T::NftCollectionId, item_id: T::NftItemId ) -> DispatchResult

pub fn do_destroy_collection( owner: T::AccountId, collection_id: T::NftCollectionId ) -> DispatchResult

impl<T: Config> Pallet<T>

pub fn create_collection( origin: OriginFor<T>, collection_id: PossibleCollections, metadata: BoundedVecOfUnq<T> ) -> DispatchResult

Creates an NFT Collection of the given Collection and sets its metadata

Parameters:

collection_id: Identifier of a Collection

metadata: Arbitrary data about a Collection, e.g. IPFS hash or name

Emits CollectionCreated event

pub fn mint( origin: OriginFor<T>, collection_id: PossibleCollections, metadata: BoundedVecOfUnq<T> ) -> DispatchResult

Mints a NFT in the specified Collection and sets its metadata

Parameters:

collection_id: The Collection of the asset to be minted.

item_id: The Collection of the asset to be minted.

metadata: Arbitrary data about an Item, e.g. IPFS hash or symbol

pub fn transfer( origin: OriginFor<T>, collection_id: PossibleCollections, item_id: T::NftItemId, dest: <T::Lookup as StaticLookup>::Source ) -> DispatchResult

Triggered by Root(origin), this transfers NFT from owner account to dest account

Parameters:

collection_id: The Collection of the asset to be transferred.

item_id: The Item of the asset to be transferred.

dest: The account to receive ownership of the asset.

pub fn burn( origin: OriginFor<T>, collection_id: PossibleCollections, item_id: T::NftItemId ) -> DispatchResult

Triggered by a servicer (origin) this removes a token from existence

Parameters:

collection_id: The Collection of the asset to be burned.

item_id: The Item of the asset to be burned.

pub fn destroy_collection( origin: OriginFor<T>, collection_id: PossibleCollections ) -> DispatchResult

Removes a Collection from existence

Parameters:

collection_id: The identifier of the asset Collection to be destroyed.

impl<T: Config> Pallet<T>

pub fn collections<KArg>(k: KArg) -> Option<CollectionInfo<BoundedVecOfUnq<T>>>where KArg: EncodeLike<T::NftCollectionId>,

Stores Collection info

impl<T: Config> Pallet<T>

pub fn items<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Option<ItemInfoOf<T>>where KArg1: EncodeLike<T::NftCollectionId>, KArg2: EncodeLike<T::NftItemId>,

Stores Item info

impl<T: Config> Pallet<T>

pub fn itemid() -> Vec<u32>

Update Item ID