Trait pallet_nft::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type ProtocolOrigin: EnsureOrigin<Self::Origin>;
    type NftClassId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::ClassId> + From<Self::ClassId>;
    type NftInstanceId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::InstanceId> + From<Self::InstanceId>;
    type ClassType: Member + Parameter + Default + Copy;
    type Permissions: NftPermission<Self::ClassType>;
    type ReserveClassIdUpTo: Get<Self::NftClassId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type WeightInfo: WeightInfo

type ProtocolOrigin: EnsureOrigin<Self::Origin>

type NftClassId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::ClassId> + From<Self::ClassId>

type NftInstanceId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::InstanceId> + From<Self::InstanceId>

type ClassType: Member + Parameter + Default + Copy

type Permissions: NftPermission<Self::ClassType>

type ReserveClassIdUpTo: Get<Self::NftClassId>

Class IDs reserved for runtime up to the following constant