Crate pallet_nft

NFT Pallet

This NFT pallet is directly inspired from the Galactic Council nft pallet. Url: https://github.com/galacticcouncil/warehouse

The NFT Pallet is used to manage & perform diverse actions on NFTs in the FairSquares framework.

Overview

Management of NFTs assets is made possible through the following actions:

NFT Collection creation

NFT asset creation or Minting

NFT asset transfer

NFT asset burning

NFT collection destruction Access to each action is restricted to a specific role.

Dispatchable Functions

create_collection - Restricted to Servicer role, this function creates an NFT Collection of the given collection and sets its metadata

mint - Restricted to Seller role, this function mints a NFT in the specified collection, and sets its metadata

transfer - Restricted to Servicer role, this function called by A(servicer) transfers NFT from account B(seller) to account C.

burn - Restricted to Servicer role, this function Removes a NFT item from existence

destroy_collection - Restricted to Servicer role, this function Removes a Collection from existence

Re-exports

pub use pallet_roles as Roles;

pub use functions::*;

pub use types::*;

pub use pallet::*;

Modules

functions

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

types

weights

Autogenerated weights for pallet_nft

Type Definitions

Acc

BoundedVecOfUnq

ItemInfoOf