Crate pallet_nft

Non-fungible tokens (NFT) processing pallet.

This creates an NFT-like pallet by implementing the Unique, Mintable, and Burnable traits of the unique_assets module.

Config

Call

Pallet

Overview

Other modules in this runtime can access the interface provided by this module to define user-facing logic to interact with the runtime NFTs.

Terminology

Usage

Interface

Supported Origins

Signed origin is valid.

Types

AssetInfo - The data type that is used to describe this type of asset. Event - Associated type for Event enum. WeightInfo - Weight information for extrinsics in this pallet.

Events

`Transferred` Event triggered when the ownership of the asset has been transferred to the account.

Errors

AssetExists\ - Thrown when there is an attempt to mint a duplicate asset. NonexistentAsset\ - Thrown when there is an attempt to transfer a nonexistent asset. NotAssetOwner\ - Thrown when someone who is not the owner of a asset attempts to transfer or burn it. DocumentNotAnchored - A given document id does not match a corresponding document in the anchor storage.

Dispatchable Functions

Callable functions (or extrinsics), also considered as transactions, materialize the pallet contract. Here’s the callable functions implemented in this module:

[transfer] - Transfer NFT [validate_mint] - Validate NFT proofs

Public Functions

Genesis Configuration

The pallet is parameterized and configured via [parameter_types] macro, at the time the runtime is built by means of the [construct_runtime] macro.

Related Pallets

This pallet is tightly coupled to the following pallets:

Substrate FRAME’s balances pallet.

Centrifuge Chain bridge pallet.

References

Substrate FRAME v2 attribute macros.

Credits

The Centrifugians Tribe tribe@centrifuge.io

License

GNU General Public License, Version 3, 29 June 2007 https://www.gnu.org/licenses/gpl-3.0.html

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

traits

Traits exported and used by non-fungible token (NFT) pallet

types

Types used by non-fungible token (NFT) pallet