Trait pallet_nft::pallet::Config

pub trait Config: Config + Config + Config + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type ChainId: Parameter + Member + Debug + Default + FullCodec + Into<u8> + From<u8>;
    type ResourceHashId: Get<ResourceId>;
    type NftProofValidationFeeKey: Get<<Self::Fees as Fees>::FeeKey>;
    type WeightInfo: WeightInfo;
}

NFT pallet’s configuration trait.

Associated types and constants are declared in this trait. If the pallet depends on other super-traits, the latter must be added to this trait, such as, in this case, [pallet_balances::Config] super-traits. Note that [frame_system::Config] must always be included.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Associated type for Event enum

type ChainId: Parameter + Member + Debug + Default + FullCodec + Into<u8> + From<u8>

Chain identifier type

type ResourceHashId: Get<ResourceId>

Resource hash id.

This type was initially declared in the bridge pallet but was moved here to avoid circular dependencies.

type NftProofValidationFeeKey: Get<<Self::Fees as Fees>::FeeKey>

Additional fee charged for validating NFT proof (when minting a NFT).

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet