Trait pallet_nft::pallet::Config

pub trait Config: Config + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type ProtocolOrigin: EnsureOrigin<Self::Origin>;
    type NftCollectionId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::CollectionId> + From<Self::CollectionId>;
    type NftItemId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::ItemId> + From<Self::ItemId>;
    type Permissions: NftPermission<Acc>;
    type ReserveCollectionIdUpTo: Get<Self::NftCollectionId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type WeightInfo: WeightInfo

type ProtocolOrigin: EnsureOrigin<Self::Origin>

type NftCollectionId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::CollectionId> + From<Self::CollectionId>

type NftItemId: Member + Parameter + Default + Copy + HasCompact + AtLeast32BitUnsigned + Into<Self::ItemId> + From<Self::ItemId>

type Permissions: NftPermission<Acc>

type ReserveCollectionIdUpTo: Get<Self::NftCollectionId>

Collection IDs reserved for runtime up to the following constant